requirements.txt
flask
geoip2
faker
python-nmap
pyopenssl